# feat: pool byte buffers during template rendering