--- a/cmd/templ/generatecmd/sourcemapvisualisation_templ.go
+++ b/cmd/templ/generatecmd/sourcemapvisualisation_templ.go
@@ -44,7 +44,8 @@ func visualisation(templFileName string, left, right templ.Component) templ.Comp
 	return templ.ComponentFunc(func(ctx context.Context, w io.Writer) (err error) {
 		templBuffer, templIsBuffer := w.(*bytes.Buffer)
 		if !templIsBuffer {
-			templBuffer = new(bytes.Buffer)
+			templBuffer = templ.GetBuffer()
+			defer templ.ReleaseBuffer(templBuffer)
 		}
 		ctx = templ.InitializeContext(ctx)
 		var_1 := templ.GetChildren(ctx)
@@ -304,7 +305,8 @@ func mappedCharacter(s string, sourceID, targetID string) templ.Component {
 	return templ.ComponentFunc(func(ctx context.Context, w io.Writer) (err error) {
 		templBuffer, templIsBuffer := w.(*bytes.Buffer)
 		if !templIsBuffer {
-			templBuffer = new(bytes.Buffer)
+			templBuffer = templ.GetBuffer()
+			defer templ.ReleaseBuffer(templBuffer)
 		}
 		ctx = templ.InitializeContext(ctx)
 		var_7 := templ.GetChildren(ctx)