--- a/generator/test-css-usage/template_templ.go
+++ b/generator/test-css-usage/template_templ.go
@@ -35,7 +35,8 @@ func Button(text string) templ.Component {
 	return templ.ComponentFunc(func(ctx context.Context, w io.Writer) (err error) {
 		templBuffer, templIsBuffer := w.(*bytes.Buffer)
 		if !templIsBuffer {
-			templBuffer = new(bytes.Buffer)
+			templBuffer = templ.GetBuffer()
+			defer templ.ReleaseBuffer(templBuffer)
 		}
 		ctx = templ.InitializeContext(ctx)
 		var_1 := templ.GetChildren(ctx)
@@ -99,7 +100,8 @@ func ThreeButtons() templ.Component {
 	return templ.ComponentFunc(func(ctx context.Context, w io.Writer) (err error) {
 		templBuffer, templIsBuffer := w.(*bytes.Buffer)
 		if !templIsBuffer {
-			templBuffer = new(bytes.Buffer)
+			templBuffer = templ.GetBuffer()
+			defer templ.ReleaseBuffer(templBuffer)
 		}
 		ctx = templ.InitializeContext(ctx)
 		var_3 := templ.GetChildren(ctx)